Move store.js file to src folder
