+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import store from './helpers/store';
+import store from './store';
 
 ReactDOM.render(
   <Provider store={store}>
